Update competitive-analysis.txt
This is a written document summarizing your competitive analysis, where you scour the web and find several existing products that are similar to what you propose. You need to use those products and carefully think about them, then list a few features in those products that are especially compelling, that you plan to include in your program. Also, list some anti-features, which are features you do not like in those products and plan to change in your program. Do not just exhaustively list all the features of those apps, since they will have tons of features that are not really relevant to your program. Stick to those few features that are directly relevant to what you are building. This is not a short piece of work. Doing a competitive analysis well will take several hours at least.